refactor(ui): split LogViewer into table/toolbar/hook
